Implementation of a wrapper for gridded data
A function named ‘gridDataWrapper’ is proposed to let the user have the
possibility to ‘transform’ an array into a proper grid data to be used
with downscaleR and the other packages.

I tested the packages a while and it seems consistent but obviously it
should be checked further.

I had to include the lubridate package to deal with monthly time
intervals (I know it’s stupid, see here for more details
http://stackoverflow.com/questions/14169620/add-a-month-to-a-date/141823
26#14182326 )